backend_test/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
pytest
pytest-django
mysqlclient
Faker
factory_boy
drf-yasg
